When play is press go to now in logs
